requirements.txt
antlr4-python3-runtime>=4.9.0
openpyxl>=3.0.0
colorama>=0.4.0
tqdm>=4.60.0